requirements.txt
pygame
paho-mqtt
flask